refactor: skeleton 및 spinner 컴포넌트 파일명 수정
